opt: Consolidate folding rules
Move all folding rules to fold_constants.opt. Move all custom functions to
a new fold_constants.go file. Consolidate tests.

Release note: None
