+++ pkg/sql/opt/norm/rules/fold_constants.opt
@@ -2,6 +2,69 @@
 # fold_constants.opt contains normalization rules for folding constants.
 # =============================================================================
 
+# FoldNullCast discards the cast operator if it has a null input. The resulting
+# null value has the same type as the Cast operator would have had.
+[FoldNullCast, Normalize]
+(Cast $input:(Null) $targetTyp:*) => (Null $targetTyp)
+
+# FoldNullUnary discards any unary operator with a null input, and replaces it
+# with a null value having the same type as the unary expression would have.
+[FoldNullUnary, Normalize]
+(Unary $input:(Null)) => (FoldNullUnary (OpName) $input)
+
+# FoldNullBinaryLeft replaces the binary operator with null if its left input
+# is null and it does not allow null arguments.
+[FoldNullBinaryLeft, Normalize]
+(Binary
+    $left:(Null)
+    $right:* & ^(AllowNullArgs (OpName) $left $right)
+)
+=>
+(FoldNullBinary (OpName) $left $right)
+
+# FoldNullBinaryRight replaces the binary operator with null if its right input
+# is null and it does not allow null arguments.
+[FoldNullBinaryRight, Normalize]
+(Binary
+    $left:*
+    $right:(Null) & ^(AllowNullArgs (OpName) $left $right)
+)
+=>
+(FoldNullBinary (OpName) $left $right)
+
+# FoldNullInNonEmpty replaces the In/NotIn with null when the left input is
+# null and the right input is not empty. Null is the unknown value, and if the
+# set is non-empty, it is unknown whether it's in/not in the set.
+[FoldNullInNonEmpty, Normalize]
+(In | NotIn
+    $left:(Null)
+    (Tuple ^[])
+)
+=>
+(Null (BoolType))
+
+# FoldNullInEmpty replaces the In with False when the left input is null and
+# the right input is empty, since even an unknown value can't be in an empty
+# set.
+[FoldNullInEmpty, Normalize]
+(In
+    $left:(Null)
+    (Tuple [])
+)
+=>
+(False)
+
+# FoldNullNotInEmpty replaces the NotIn with True when the left input is null
+# and the right input is empty, since even an unknown value can't be in an
+# empty set.
+[FoldNullNotInEmpty, Normalize]
+(NotIn
+    $left:(Null)
+    (Tuple [])
+)
+=>
+(True)
+
 # FoldArray evaluates an Array expression with constant inputs. It replaces the
 # Array with a Const datum with type TArray.
 [FoldArray, Normalize]
@@ -22,10 +85,7 @@
 #
 # In this query, the ELSE clause is not executed, so the divide-by-zero error
 # should not be triggered.
-#
-# This rule is marked as low priority so that the FoldNull rules in scalar.opt
-# can run first.
-[FoldBinary, Normalize, LowPriority]
+[FoldBinary, Normalize]
 (Binary
     $left:* & (IsConstValueOrTuple $left)
     $right:* &
@@ -38,10 +98,7 @@ $result
 # FoldUnary is similar to FoldBinary, but it involves a unary operation over a
 # single constant input. As with FoldBinary, FoldUnary applies as long as the
 # evaluation would not cause an error.
-#
-# This rule is marked as low priority so that the FoldNull rules in scalar.opt
-# can run first.
-[FoldUnary, Normalize, LowPriority]
+[FoldUnary, Normalize]
 (Unary
     $input:* &
         (IsConstValueOrTuple $input) &
@@ -53,10 +110,7 @@ $result
 # FoldComparison is similar to FoldBinary, but it involves a comparison
 # operation. As with FoldBinary, FoldComparison applies as long as the
 # evaluation would not cause an error.
-#
-# This rule is marked as low priority so that the FoldNull rules in scalar.opt
-# can run first.
-[FoldComparison, Normalize, LowPriority]
+[FoldComparison, Normalize]
 (Comparison
     $left:* & (IsConstValueOrTuple $left)
     $right:* &
@@ -69,10 +123,7 @@ $result
 # FoldCast is similar to FoldUnary, but it involves a cast operation. As with
 # FoldUnary, FoldCast applies as long as the evaluation would not cause an
 # error.
-#
-# This rule is marked as low priority so that the FoldNull rules in scalar.opt
-# can run first.
-[FoldCast, Normalize, LowPriority]
+[FoldCast, Normalize]
 (Cast
     $input:*
     $typ:* &
@@ -87,13 +138,10 @@ $result
 # evaluation would not cause an error. Additionally, only certain functions
 # are safe to fold as part of normalization. Other functions rely on context
 # that may change between runs of a prepared query.
-#
-# This rule is marked as low priority so that the FoldNull rules in scalar.opt
-# can run first.
-[FoldFunction, Normalize, LowPriority]
+[FoldFunction, Normalize]
 (Function
-  $args:* & (IsListOfConstants $args)
-  $private:* & (Succeeded $result:(FoldFunction $args $private))
+    $args:* & (IsListOfConstants $args)
+    $private:* & (Succeeded $result:(FoldFunction $args $private))
 )
 =>
 $result
