+++ pkg/sql/opt/norm/rules/scalar.opt
@@ -55,69 +55,6 @@
 [EliminateCast, Normalize]
 (Cast $input:* $targetTyp:* & (HasColType $input $targetTyp)) => $input
 
-# FoldNullCast discards the cast operator if it has a null input. The resulting
-# null value has the same type as the Cast operator would have had.
-[FoldNullCast, Normalize]
-(Cast $input:(Null) $targetTyp:*) => (Null $targetTyp)
-
-# FoldNullUnary discards any unary operator with a null input, and replaces it
-# with a null value having the same type as the unary expression would have.
-[FoldNullUnary, Normalize]
-(Unary $input:(Null)) => (FoldNullUnary (OpName) $input)
-
-# FoldNullBinaryLeft replaces the binary operator with null if its left input
-# is null and it does not allow null arguments.
-[FoldNullBinaryLeft, Normalize]
-(Binary
-    $left:(Null)
-    $right:* & ^(AllowNullArgs (OpName) $left $right)
-)
-=>
-(FoldNullBinary (OpName) $left $right)
-
-# FoldNullBinaryRight replaces the binary operator with null if its right input
-# is null and it does not allow null arguments.
-[FoldNullBinaryRight, Normalize]
-(Binary
-    $left:*
-    $right:(Null) & ^(AllowNullArgs (OpName) $left $right)
-)
-=>
-(FoldNullBinary (OpName) $left $right)
-
-# FoldNullInNonEmpty replaces the In/NotIn with null when the left input is
-# null and the right input is not empty. Null is the unknown value, and if the
-# set is non-empty, it is unknown whether it's in/not in the set.
-[FoldNullInNonEmpty, Normalize]
-(In | NotIn
-    $left:(Null)
-    (Tuple ^[])
-)
-=>
-(Null (BoolType))
-
-# FoldNullInEmpty replaces the In with False when the left input is null and
-# the right input is empty, since even an unknown value can't be in an empty
-# set.
-[FoldNullInEmpty, Normalize]
-(In
-    $left:(Null)
-    (Tuple [])
-)
-=>
-(False)
-
-# FoldNullNotInEmpty replaces the NotIn with True when the left input is null
-# and the right input is empty, since even an unknown value can't be in an
-# empty set.
-[FoldNullNotInEmpty, Normalize]
-(NotIn
-    $left:(Null)
-    (Tuple [])
-)
-=>
-(True)
-
 # NormalizeInConst ensures that the In operator's tuple operand is sorted with
 # duplicates removed (since duplicates do not change the result).
 [NormalizeInConst, Normalize]
