+++ pkg/sql/opt/norm/testdata/rules/scalar
@@ -199,129 +199,6 @@ project
       └── ARRAY[a.i, 2]::INT2VECTOR [type=int2vector, outer=(2)]
 
 # --------------------------------------------------
-# FoldNullCast
-# --------------------------------------------------
-opt expect=FoldNullCast
-SELECT
-    null::int,
-    null::timestamptz,
-    null::decimal(19,2)::int::bit::char(2),
-    null::oidvector,
-    null::int2vector
-----
-values
- ├── columns: int8:1(int) timestamptz:2(timestamptz) bpchar:3(char) oid:4(oidvector) int2:5(int2vector)
- ├── cardinality: [1 - 1]
- ├── key: ()
- ├── fd: ()-->(1-5)
- └── (NULL, NULL, NULL, NULL, NULL) [type=tuple{int, timestamptz, char, oidvector, int2vector}]
-
-# --------------------------------------------------
-# FoldNullUnary
-# --------------------------------------------------
-opt expect=FoldNullUnary
-SELECT +null::int AS r, -null::int AS s, ~null::int AS t FROM a
-----
-project
- ├── columns: r:7(int) s:8(int) t:9(int)
- ├── fd: ()-->(7-9)
- ├── scan a
- └── projections
-      ├── null [type=int]
-      ├── null [type=int]
-      └── null [type=int]
-
-# --------------------------------------------------
-# FoldNullBinaryLeft, FoldNullBinaryRight
-# --------------------------------------------------
-opt expect=(FoldNullBinaryLeft,FoldNullBinaryRight)
-SELECT
-    null::int & 1 AS ra, 1 & null::int AS rb,
-    null::decimal + 1 AS sa, 1 + null::decimal AS sb,
-    null::float % 1 AS ta, 1 % null::float AS tb,
-    null::int << 4 AS ua, 4 << null::int AS ub,
-
-    -- These shouldn't be folded because AllowNullArgs is true for concat with arrays.
-    arr::decimal[] || null AS va, null || arr::string[] AS vb,
-
-    -- Scalars concatenated with nulls match array overloads, and shouldn't be folded.
-    -- In other words, the only overload for decimal concatenation is an array overload.
-    i::decimal || null AS wa, null || i::float AS wb
-FROM a
-----
-project
- ├── columns: ra:7(int) rb:8(int) sa:9(decimal) sb:10(decimal) ta:11(float) tb:12(float) ua:13(int) ub:14(int) va:15(decimal[]) vb:16(string[]) wa:17(decimal[]) wb:18(float[])
- ├── fd: ()-->(7-14)
- ├── scan a
- │    └── columns: i:2(int) arr:6(int[])
- └── projections
-      ├── null [type=int]
-      ├── null [type=int]
-      ├── null [type=decimal]
-      ├── null [type=decimal]
-      ├── null [type=float]
-      ├── null [type=float]
-      ├── null [type=int]
-      ├── null [type=int]
-      ├── arr::DECIMAL[] || CAST(NULL AS DECIMAL[]) [type=decimal[], outer=(6)]
-      ├── CAST(NULL AS STRING[]) || arr::STRING[] [type=string[], outer=(6)]
-      ├── i::DECIMAL || CAST(NULL AS DECIMAL[]) [type=decimal[], outer=(2)]
-      └── CAST(NULL AS FLOAT8[]) || i::FLOAT8 [type=float[], outer=(2)]
-
-opt
-SELECT
-    null::json || '[1, 2]' AS ra, '[1, 2]' || null::json AS rb,
-    null::json->'foo' AS sa, '{}'::jsonb->null::string AS sb,
-    null::json->>'foo' AS ta, '{}'::jsonb->>null::string AS tb,
-    null::json->>'foo' AS ua, '{}'::jsonb->>null::string AS ub,
-    null::json#>ARRAY['foo'] AS va, '{}'::jsonb#>NULL AS vb,
-    null::json#>>ARRAY['foo'] AS wa, '{}'::jsonb#>>NULL AS wb
-FROM a
-----
-project
- ├── columns: ra:7(jsonb) rb:8(jsonb) sa:9(jsonb) sb:10(jsonb) ta:11(string) tb:12(string) ua:11(string) ub:12(string) va:13(jsonb) vb:14(unknown) wa:15(string) wb:14(unknown)
- ├── fd: ()-->(7-15)
- ├── scan a
- └── projections
-      ├── null [type=jsonb]
-      ├── null [type=jsonb]
-      ├── null [type=jsonb]
-      ├── null [type=jsonb]
-      ├── null [type=string]
-      ├── null [type=string]
-      ├── null [type=jsonb]
-      ├── null [type=unknown]
-      └── null [type=string]
-
-# --------------------------------------------------
-# FoldNullInNonEmpty
-# --------------------------------------------------
-opt expect=FoldNullInNonEmpty
-SELECT null IN (i) AS r, null NOT IN (s) AS s FROM a
-----
-project
- ├── columns: r:7(bool) s:8(bool)
- ├── fd: ()-->(7,8)
- ├── scan a
- └── projections
-      ├── null [type=bool]
-      └── null [type=bool]
-
-# --------------------------------------------------
-# FoldInNull
-# --------------------------------------------------
-opt expect=FoldInNull
-SELECT i IN (null, null) AS r, k NOT IN (1 * null, null::int, 1 < null) AS s FROM a
-----
-project
- ├── columns: r:7(bool) s:8(bool)
- ├── fd: ()-->(7,8)
- ├── scan a
- └── projections
-      ├── null [type=bool]
-      └── null [type=bool]
-
-# --------------------------------------------------
 # NormalizeInConst
 # --------------------------------------------------
 opt expect=NormalizeInConst
